requirements.txt
alpaca-py>=0.40.0
pandas>=1.0.0
numpy>=1.18.0
python-dotenv>=0.15.0
requests>=2.25.0
pytest>=7.0.0
tenacity>=8.0.0
apscheduler>=3.9.1 # for scheduling live runs
python-json-logger>=2.0.0
joblib>=1.0.0
backtrader>=1.9.74.123
scikit-learn>=1.0.0
textblob>=0.17.1
matplotlib>=3.0.0
